ML-Service/requirements.txt
torch~=2.8.0
torchvision~=0.23.0
numpy~=2.2.6
ultralytics~=8.3.203
requests~=2.32.5
pillow~=11.3.0
fastapi~=0.118.0
uvicorn~=0.37.0